bug solved in update policy

+++ src/main/java/com/beercenter/shop/core/services/StockFileProcessorServiceImpl.java
@@ -4,6 +4,7 @@ import com.beercenter.shop.core.beans.ConfigProperties;
 import com.beercenter.shop.core.model.InventoryLevel;
 import com.beercenter.shop.core.model.ProductStockInfo;
 import com.beercenter.shop.core.model.Variant;
+import com.beercenter.shop.core.utils.InventoryPolicy;
 import com.opencsv.bean.CsvToBeanBuilder;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -16,8 +17,7 @@ import java.nio.file.Path;
 import java.util.*;
 import java.util.stream.Collectors;
 
-import static com.beercenter.shop.core.utils.InventoryPolicy.ONLY_AT_STORE;
-import static com.beercenter.shop.core.utils.InventoryPolicy.REGULAR;
+import static com.beercenter.shop.core.utils.InventoryPolicy.*;
 
 @AllArgsConstructor
 @Slf4j
@@ -30,7 +30,7 @@ public class StockFileProcessorServiceImpl {
     private StringBuilder stringBuilder;
     private Set<Variant> variantUpdateFails;
 
-    public void processFile(final Path filePath) throws FileNotFoundException {
+    public void processFile(final Path filePath) {
         cleanVariables();
         try {
             log.info("START -> Execution");
@@ -38,8 +38,8 @@ public class StockFileProcessorServiceImpl {
             final Map<String, Long> productsStockMap = getProductsStockMapFromFile(filePath);
             final List<Variant> variantsToUpdate = getVariantsToUpdate(variants, productsStockMap);
             if (!CollectionUtils.isEmpty(variantsToUpdate)) {
-                updateProductsInventoryPolicy(variantsToUpdate);
-                updateProductsStock(variantsToUpdate);
+                final Map<Long, InventoryLevel> inventoryLevelMap = productServiceImpl.getVariantInventoryList(variantsToUpdate);
+                variantsToUpdate.forEach(variant -> updateProduct(variant, inventoryLevelMap));
             }
             log.info(String.format("END -> Execution { TOTAL READ : %s , TOTAL UPDATED : %s }", productsStockMap.size(), variantsToUpdate.size()));
             addInfo("TOTAL PRODUCTS READ: " + productsStockMap.size());
@@ -50,6 +50,19 @@ public class StockFileProcessorServiceImpl {
         } catch (final Exception e) {
             addInfo("ERROR: " + ExceptionUtils.getMessage(e));
             log.error("Error has occurred while file processing: ", e);
+        }
+    }
+
+    private void updateProduct(final Variant variant, final Map<Long, InventoryLevel> inventoryLevelMap) {
+        try {
+            log.info("UPDATING PRODUCT: " + variant.getSku());
+            productServiceImpl.updateProductVariant(Variant.builder().inventory_policy(variant.getInventory_policy()).inventory_management(variant.getInventory_management()).id(variant.getId()).build());
+            productServiceImpl.adjustVariantInventory(inventoryLevelMap.get(variant.getInventory_item_id()), variant.getStockAdjust());
+        } catch (Exception e) {
+            final String message = "ERROR UPDATING PRODUCT " + variant.getSku() + " ERROR: " + ExceptionUtils.getMessage(e);
+            log.info(message);
+            addInfo(message);
+            variantUpdateFails.add(variant);
         }
     }
 
@@ -84,38 +97,6 @@ public class StockFileProcessorServiceImpl {
         return variants;
     }
 
-    private void updateProductsInventoryPolicy(final List<Variant> variantsToUpdate) {
-        log.info("START -> Updating products policy");
-        variantsToUpdate.forEach(variant -> {
-            try {
-                productServiceImpl.updateProductVariant(Variant.builder().inventory_policy(variant.getInventory_policy()).id(variant.getId()).build());
-            } catch (Exception e) {
-                final String message = "ERROR UPDATING POLICY OF PRODUCT " + variant.getSku();
-                log.info(message);
-                addInfo(message);
-                variantUpdateFails.add(variant);
-            }
-        });
-        log.info("END -> Updating products policy");
-    }
-
-    private void updateProductsStock(final List<Variant> variantsToUpdate) {
-        log.info("START -> Updating products stock");
-        final Map<Long, InventoryLevel> inventoryLevelMap = productServiceImpl.getVariantInventoryList(variantsToUpdate);
-
-        variantsToUpdate.forEach(variant -> {
-            try {
-                productServiceImpl.adjustVariantInventory(inventoryLevelMap.get(variant.getInventory_item_id()), variant.getStockAdjust());
-            } catch (Exception e) {
-                final String message = "ERROR UPDATING STOCK OF PRODUCT " + variant.getSku();
-                log.info(message);
-                addInfo(message);
-                variantUpdateFails.add(variant);
-            }
-        });
-        log.info("END -> Updating products stock");
-    }
-
     private List<Variant> getVariantsToUpdate(Set<Variant> variants, Map<String, Long> productsStockMap) {
         final List<Variant> variantsToUpdate = new ArrayList<>();
         log.info("START -> Get products that need to be updated");
@@ -124,8 +105,10 @@ public class StockFileProcessorServiceImpl {
             if (stock != null) {
                 if (variant.getInventory_quantity().compareTo(stock) != 0) {
                     if (stock > 2 || stock <= 0) {
+                        variant.setInventory_management(MANAGEMENT_SHOPIFY.getValue());
                         variant.setInventory_policy(REGULAR.getValue());
                     } else {
+                        variant.setInventory_management(MANAGEMENT_SHOPIFY.getValue());
                         variant.setInventory_policy(ONLY_AT_STORE.getValue());
                     }
                     variant.setStockAdjust(stock - variant.getInventory_quantity());
@@ -151,13 +134,24 @@ public class StockFileProcessorServiceImpl {
                     .withType(ProductStockInfo.class).withSeparator(';').build().parse();
             log.info("END -> Reading products from file");
         } catch (final Exception e) {
-            fileReader.close();
-            stockFile.delete();
+            closeFileReader(fileReader);
+            deleteFile(stockFile);
             throw e;
         }
-        fileReader.close();
-        stockFile.delete();
+        closeFileReader(fileReader);
+        deleteFile(stockFile);
 
         return productStockInfoList.stream().collect(Collectors.toMap(ProductStockInfo::getSku, ProductStockInfo::getStock));
     }
+
+    private void deleteFile(final File stockFile) {
+        if (stockFile != null)
+        stockFile.delete();
+    }
+
+    private void closeFileReader(final FileReader fileReader) throws IOException {
+        if (fileReader != null)
+        fileReader.close();
+    }
 }
+
